feat: simple example that uses example syntax that is compatible with the SDK documentation as of 1.6.0
I added this example to test and debug with something simple. It is similar to the everything example but it is more simple and uses the syntax for adding resources and tools that is shown in the SDK docs instead of 'setRequestHandler'.
Sharing this in case it might be helpful to others