feat: expose GitHub server instance to npm library
This commit separates the `main` logic into a separate file, now marked
as the binary entry point for the package, and updates the index to
export the server instance. This allows the package to be used as a
library, enabling the server to be used with custom transports in other
projects.